ml/requirements.txt
google-cloud-aiplatform>=1.40.0
google-cloud-bigquery>=3.17.0
google-cloud-storage>=2.14.0
scikit-learn>=1.4.0
pandas>=2.2.0
python-dotenv>=1.0.0
